Add AWS Lambda export option for all nodes; move export logic to helper functions.
