+++ web/main/templates/includes/action_buttons.html
@@ -53,7 +53,7 @@
     {% if not resource and not section and request.user.is_superuser and not FORCE_EXPERIMENTAL_EXPORT %}
       <a class="action one-line export export-experimental export-{{ section.is_annotated|yesno:"has,no" }}-annotations" role="button" rel="nofollow" href="#">Experimental Export</a>
     {% endif %}
-    {% if not resource and not section and request.user.is_superuser and not FORCE_AWS_LAMBDA_EXPORT %}
+    {% if request.user.is_superuser and not FORCE_AWS_LAMBDA_EXPORT %}
       <a class="action one-line export export-lambda export-{{ section.is_annotated|yesno:"has,no" }}-annotations" role="button" rel="nofollow" href="#">AWS Lambda Export</a>
     {% endif %}
   {% endif %}
